+++ controllers/contacts/index.js
@@ -1,0 +1,16 @@
+const listContacts = require("./listContacts");
+const addContact = require("./addContact");
+const getContactById = require("./getContactById");
+const removeContact = require("./removeContact");
+const updateContact = require("./updateContact");
+const updateStatusContact = require("./updateStatusContact");
+const { controllerWrapper } = require("../../decorators/controllerWrapper");
+
+module.exports = {
+  listContacts: controllerWrapper(listContacts),
+  getContactById: controllerWrapper(getContactById),
+  addContact: controllerWrapper(addContact),
+  removeContact: controllerWrapper(removeContact),
+  updateContact: controllerWrapper(updateContact),
+  updateStatusContact: controllerWrapper(updateStatusContact),
+};
