+++ controllers/contacts/removeContact.js
@@ -1,0 +1,16 @@
+const { HttpError } = require("../../helpers");
+const contactsService = require("../../models");
+
+const removeContact = async (req, res) => {
+  const { contactId } = req.params;
+
+  const result = await contactsService.removeContact(contactId);
+  if (!result) {
+    throw HttpError(404, `Contact with ${contactId} not found`);
+  }
+  res.status(200).json({
+    message: "Delete success",
+  });
+};
+
+module.exports = removeContact;
