+++ routes/api/contacts.js
@@ -2,7 +2,7 @@ const express = require("express");
 
 const router = express.Router();
 
-const contactsController = require("../../controllers/contacts-controller");
+const contactsController = require("../../controllers/contacts");
 
 router.get("/", contactsController.listContacts);
 
@@ -17,3 +17,5 @@ router.put("/:contactId", contactsController.updateContact);
 router.patch("/:contactId/favorite", contactsController.updateStatusContact);
 
 module.exports = router;
+
+//PATCH  - http://localhost:3000/api/contacts/6468fdfea8612f4344b19479/favorite
